first cut at calculator service with tests that don't break

+++ math-functions-lib/src/main/java/org/example/Subtractor.java
@@ -2,7 +2,7 @@ package main.java.org.example;
 
 public class Subtractor {
 
-    public int subtract( int i, int j ) {
+    public double subtract( double i, double j ) {
         return i - j;
     }
